refactor(history): split queries TU out of shothistorystorage.cpp (J pt 2)
Second half of openspec/changes/split-shothistorystorage-by-concern/.
Pure refactor — no public API change, no behaviour change. After PR
#951 (serialize split) the main TU was still ~3700 lines because the
query / read-projection code lived alongside DB lifecycle, save, and
load+recompute. This PR moves that block into its own TU.

Moves to src/history/shothistorystorage_queries.cpp (~1140 lines):
- Filtered list: requestShotsFiltered, parseFilterMap, buildFilterQuery,
  formatFtsQuery, s_sortColumnMap whitelist.
- Recents-by-kbId: requestRecentShotsByKbId, loadRecentShotsByKbIdStatic.
- Distinct-value cache: requestDistinctCache, requestDistinctValueAsync,
  getDistinctValues, invalidateDistinctCache, s_allowedColumns whitelist,
  the getDistinct{BeanBrands,BeanTypes,Grinders,GrinderSettings,Baristas,
  BeanTypesForBrand,GrinderBrands,GrinderModelsForBrand,
  GrinderBurrsForModel,GrinderSettingsForGrinder} family, and
  sortGrinderSettings (only called from requestDistinctValueAsync).
- Auto-favorites: requestAutoFavorites, requestAutoFavoriteGroupDetails.
- Grinder context: queryGrinderContext, requestUpdateGrinderFields.

The queries TU shares the use12h() helper from shothistorystorage_internal.h
(via a TU-scope using directive); no other internal helpers needed.

Stays in shothistorystorage.cpp:
- DB lifecycle (initialize, createTables, runMigrations, close)
- Save path (saveShot, saveShotStatic, compressSampleData, etc.)
- Load + recompute (requestShot, requestReanalyzeBadges,
  loadShotRecordStatic, computeDerivedCurves, computePhaseSummaries)
- Delete + update (deleteShot, deleteShots, updateShotMetadataStatic,
  requestUpdateShotMetadata, requestDeleteShot)
- Backup/import (createBackupStatic, importDatabaseStatic,
  performDatabaseCopy, requestCreateBackup, requestImportDatabase,
  importShotRecord, getShotCountStatic)
- Misc lifecycle helpers (updateTotalShots, refreshTotalShots,
  backfillBeverageType, checkpoint)

CMake: root CMakeLists.txt and tests/CMakeLists.txt updated. No header
change for the queries TU — its contents are member definitions on the
existing ShotHistoryStorage class.

Build clean via Qt Creator MCP. 1811 tests pass, 0 warnings emitted by
tests (the 2 build-time warnings are pre-existing OpenSSL link-version
mismatches in the dev environment).

Line counts post-J:
  shothistorystorage.cpp           2584  (was 3925; -1341)
  shothistorystorage_queries.cpp   1139  (new in this PR)
  shothistorystorage_serialize.cpp  215  (split in PR #951)
  shothistorystorage_internal.cpp    47  (split in PR #951)
  shothistorystorage_internal.h      52  (split in PR #951)
