+++ openspec/changes/split-shothistorystorage-by-concern/tasks.md
@@ -2,12 +2,12 @@
 
 ## 1. Queries split (~700 lines)
 
-- [ ] 1.1 Create `src/history/shothistorystorage_queries.cpp`. Add the same `#include` block the main TU uses (subset for query needs).
-- [ ] 1.2 Move `requestShotsFiltered`, the `buildFilterQuery` static helper, the QSqlQuery binding code into the new file.
-- [ ] 1.3 Move `requestRecentShotsByKbId`, `loadRecentShotsByKbIdStatic`.
-- [ ] 1.4 Move `requestAutoFavorites`, `requestAutoFavoriteGroupDetails`, the auto-favorite SQL templates.
-- [ ] 1.5 Move `requestDistinctCache`, the distinct-value cache mutation, `getDistinctBeanBrands` and friends, `getDistinctBeanTypesForBrand`, `getDistinctGrinderBrands`, etc.
-- [ ] 1.6 Move `queryGrinderContext`, `requestUpdateGrinderFields`.
+- [x] 1.1 Created `src/history/shothistorystorage_queries.cpp` with a subset `#include` block (Qt SQL/json/locale + `core/dbutils.h` + `core/grinderaliases.h`) plus a `using decenza::storage::detail::use12h;` for the shared helper. Header preamble enumerates the five owning concerns (filtered queries, recents-by-kbId, distinct cache, auto-favorites, grinder context).
+- [x] 1.2 Moved `requestShotsFiltered`, `buildFilterQuery`, `parseFilterMap`, `formatFtsQuery`, and the `s_sortColumnMap` whitelist into the new file. The QSqlQuery binding lambda inside `requestShotsFiltered` moves with it.
+- [x] 1.3 Moved `requestRecentShotsByKbId` and `loadRecentShotsByKbIdStatic`.
+- [x] 1.4 Moved `requestAutoFavorites` and `requestAutoFavoriteGroupDetails`. The auto-favorite SQL templates and the `use12h()` call inside the latter move along with them.
+- [x] 1.5 Moved `requestDistinctCache`, `requestDistinctValueAsync`, `getDistinctValues`, `invalidateDistinctCache`, the `s_allowedColumns` whitelist, and the `getDistinct{BeanBrands,BeanTypes,Grinders,GrinderSettings,Baristas,BeanTypesForBrand,GrinderBrands,GrinderModelsForBrand,GrinderBurrsForModel,GrinderSettingsForGrinder}` family. `sortGrinderSettings` moves with them since it's only called from `requestDistinctValueAsync`.
+- [x] 1.6 Moved `queryGrinderContext` and `requestUpdateGrinderFields`.
 
 ## 2. Serialization split (~200 lines)
 
@@ -16,16 +16,16 @@
 
 ## 3. CMake
 
-- [x] 3.1 Updated `CMakeLists.txt` (root): `src/history/shothistorystorage_internal.cpp` and `src/history/shothistorystorage_serialize.cpp` added alongside `shothistorystorage.cpp`. `shothistorystorage_internal.h` added to the header list (parallel to `shothistorystorage.h`).
-- [x] 3.2 Updated `tests/CMakeLists.txt`: `HISTORY_SOURCES` and the two MCP test targets that pull in `shothistorystorage.cpp` directly (`tst_mcptools_write` at line 478, `tst_mcpserver_session` at line 502) all picked up the two new TUs via a `replace_all` edit on the source line.
+- [x] 3.1 Updated `CMakeLists.txt` (root): all three new TUs (`shothistorystorage_internal.cpp`, `shothistorystorage_serialize.cpp`, `shothistorystorage_queries.cpp`) are listed alongside `shothistorystorage.cpp`. `shothistorystorage_internal.h` added to the header list (parallel to `shothistorystorage.h`). Header for the queries TU is intentionally omitted — its functions are member definitions on `ShotHistoryStorage` and need no separate header.
+- [x] 3.2 Updated `tests/CMakeLists.txt`: `HISTORY_SOURCES` and the two MCP test targets that pull in `shothistorystorage.cpp` directly (`tst_mcptools_write`, `tst_mcpserver_session`) all picked up the queries TU via the same `replace_all` mechanism that added the serialize TU in part 1.
 
 ## 4. Verify
 
 - [x] 4.1 Build clean via Qt Creator MCP — succeeded with 0 errors. The 2 warnings reported (`linking with dylib '...openssl@3...' which was built for newer version 26.0`) are pre-existing OpenSSL-version mismatches in the dev environment, not introduced by this PR.
 - [x] 4.2 All 1811 tests pass; 0 failed, 0 with warnings.
 - [x] 4.3 Re-grepped for `convertShotRecord` / `prepareAnalysisInputs` / `profileFrameInfoFromJson` / `use12h` — each function has exactly one definition; matches in other files are declarations, comments, or `using` aliases.
-- [x] 4.4 Line counts post-split: `shothistorystorage.cpp` = 3679 (down from 3925), `shothistorystorage_serialize.cpp` = 214, `shothistorystorage_internal.cpp` = 47, `shothistorystorage_internal.h` = 51. The serialize split shed ~250 lines from the main TU; the queries split (1.x, deferred to a follow-up) will trim it further.
+- [x] 4.4 Line counts post-split: `shothistorystorage.cpp` = 2584 (down from 3925), `shothistorystorage_queries.cpp` = 1139, `shothistorystorage_serialize.cpp` = 215, `shothistorystorage_internal.cpp` = 47, `shothistorystorage_internal.h` = 52. Combined: the four TUs sum to 4037 lines (33% larger than the original; growth comes from the per-TU includes/preambles), but no single file is now larger than 2584. Main TU shed ~1300 lines across the two splits.
 
 ## 5. Optional
 
-- [x] 5.1 Per the proposal's recommendation, this PR ships ONLY the serialization split + the shared-internal-helpers extraction. The queries split (tasks 1.1–1.6) is deferred to a follow-up PR — easier to review, smaller diff, and the queries body is independent of the serialize move so the order doesn't matter.
+- [x] 5.1 Shipped as two PRs per the proposal's recommendation: PR 1 (`#951`) carried the serialization split + shared internal helpers; PR 2 (this one) carries the queries split. The queries body was independent of the serialize move so the two PRs landed in either order.
